+++ apps/ticket_web/lib/pages/home/components/ticket_cards/personal_sponsor_ticket_card.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_markdown/flutter_markdown.dart';
 import 'package:intl/intl.dart';
 import 'package:ticket_web/gen/i18n/strings.g.dart';
 
@@ -46,11 +47,14 @@ class PersonalSponsorTicketCard extends StatelessWidget {
               ),
             ),
             const SizedBox(height: 4),
-            Text(
-              i18n.homePage.tickets.personalSponsor.description,
-              style: textTheme.bodyLarge?.copyWith(
-                color: theme.colorScheme.onSurface.withValues(alpha: 0.7),
+            MarkdownBody(
+              data: i18n.homePage.tickets.personalSponsor.description,
+              styleSheet: MarkdownStyleSheet.fromTheme(theme).copyWith(
+                p: textTheme.bodyLarge?.copyWith(
+                  color: theme.colorScheme.onSurface.withValues(alpha: 0.7),
+                ),
               ),
+              softLineBreak: true,
             ),
             const SizedBox(height: 16),
             FilledButton.icon(
